+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## Unreleased
+* Add AuthorizationAdjustment class and `authorizationAdjustments` to Transaction
+
+## Unreleased
 * Add `IDEAL_PAYMENT` to `PaymentInstrumentType`
 * Create Braintree class to be PSR compliant